Remove Addins infrastructure
