--- src/framework/Extensibility/AddinStatus.cs
@@ -1,56 +1,0 @@
-// ***********************************************************************
-// Copyright (c) 2007 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-#if !NUNITLITE
-using System;
-
-namespace NUnit.Framework.Extensibility
-{
-	/// <summary>
-	/// The AddinStatus enum indicates the load status of an addin.
-	/// </summary>
-    public enum AddinStatus
-    {
-		/// <summary>
-		/// Not known - default
-		/// </summary>
-        Unknown,
-		/// <summary>
-		/// The addin is enabled but not loaded
-		/// </summary>
-        Enabled,
-		/// <summary>
-		/// The addin is disabled
-		/// </summary>
-        Disabled,
-		/// <summary>
-		/// The addin was loaded successfully
-		/// </summary>
-        Loaded,
-		/// <summary>
-		/// An error was encountered loading the addin
-		/// </summary>
-        Error
-    }
-}
-#endif
