--- src/framework/Extensibility/IAddinRegistry.cs
@@ -1,63 +1,0 @@
-// ***********************************************************************
-// Copyright (c) 2007 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-#if !NUNITLITE
-using System;
-
-namespace NUnit.Framework.Extensibility
-{
-	/// <summary>
-	/// The IAddinRegistry interface allows registering addins
-	/// and retrieving information about them. It is also used
-	///  to record the load status of an addin.
-	/// </summary>
-	public interface IAddinRegistry
-	{
-		/// <summary>
-		/// Gets a list of all addins as Addin objects
-		/// </summary>
-		System.Collections.IList Addins { get; }
-
-		/// <summary>
-		/// Registers an addin
-		/// </summary>
-		/// <param name="addin">The addin to be registered</param>
-		void Register( Addin addin );
-
-        /// <summary>
-        /// Returns true if an addin of a given name is registered
-        /// </summary>
-        /// <param name="name">The name of the addin</param>
-        /// <returns>True if an addin of that name is registered, otherwise false</returns>
-        bool IsAddinRegistered(string name);
-
-		/// <summary>
-		///  Sets the load status of an addin
-		/// </summary>
-		/// <param name="name">The name of the addin</param>
-		/// <param name="status">The status to be set</param>
-		/// <param name="message">An optional message explaining the status</param>
-		void SetStatus( string name, AddinStatus status, string message );
-	}
-}
-#endif
