--- src/framework/Extensibility/NUnitAddinAttribute.cs
@@ -1,60 +1,0 @@
-// ***********************************************************************
-// Copyright (c) 2007 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-#if !NUNITLITE
-using System;
-
-namespace NUnit.Framework.Extensibility
-{
-	/// <summary>
-	/// NUnitAddinAttribute is used to mark all add-ins. The marked class
-	/// must implement the IAddin interface.
-	/// </summary>
-	[AttributeUsage(AttributeTargets.Class, AllowMultiple=false, Inherited=false)]
-	public sealed class NUnitAddinAttribute : Attribute
-	{
-        /// <summary>
-        /// The name of this addin
-        /// </summary>
-		public string Name;
-
-        /// <summary>
-        /// A description for the addin
-        /// </summary>
-		public string Description;
-
-        /// <summary>
-        /// The type of extension provided
-        /// </summary>
-		public ExtensionType Type;
-
-        /// <summary>
-        /// Default Constructor
-        /// </summary>
-		public NUnitAddinAttribute()
-		{
-			this.Type = ExtensionType.Core;
-		}
-	}
-}
-#endif
