--- src/framework/Extensibility/TestFramework.cs
@@ -1,62 +1,0 @@
-// ***********************************************************************
-// Copyright (c) 2007 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-#if !NUNITLITE
-using System;
-
-namespace NUnit.Framework.Extensibility
-{
-	/// <summary>
-	/// Summary description for TestFramework.
-	/// </summary>
-	[Serializable]
-	public class TestFramework
-	{
-		#region Instance Fields
-		/// <summary>
-		/// The name of the framework
-		/// </summary>
-		public string Name;
-
-		/// <summary>
-		/// The file name of the assembly that defines the framwork
-		/// </summary>
-		public string AssemblyName;
-		#endregion
-
-		#region Constructor
-		/// <summary>
-		/// Constructs a TestFramwork object given its name and assembly name.
-		/// </summary>
-		/// <param name="frameworkName"></param>
-		/// <param name="assemblyName"></param>
-		public TestFramework( string frameworkName, string assemblyName ) 
-		{
-			this.Name = frameworkName;
-			this.AssemblyName = assemblyName;
-		}
-		#endregion
-	}
-}
-#endif
-
