+++ src/framework/Internal/Builders/DataAttributeTestCaseProvider.cs
@@ -26,7 +26,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
 using NUnit.Framework.Interfaces;
-using NUnit.Framework.Extensibility;
 
 namespace NUnit.Framework.Internal.Builders
 {
