+++ src/framework/Internal/Builders/LegacySuiteBuilder.cs
@@ -26,7 +26,6 @@ using System;
 using System.Collections;
 using System.Reflection;
 using NUnit.Framework.Interfaces;
-using NUnit.Framework.Extensibility;
 
 namespace NUnit.Framework.Internal.Builders
 {
