+++ src/framework/Internal/Builders/PairwiseStrategy.cs
@@ -27,7 +27,6 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 using NUnit.Framework.Interfaces;
-using NUnit.Framework.Extensibility;
 
 namespace NUnit.Framework.Internal.Builders
 {
