+++ src/framework/Internal/Builders/SetUpFixtureBuilder.cs
@@ -23,7 +23,6 @@
 
 using System;
 using NUnit.Framework.Interfaces;
-using NUnit.Framework.Extensibility;
 
 namespace NUnit.Framework.Internal.Builders
 {
