--- src/framework/Internal/Extensibility/SuiteBuilderCollection.cs
@@ -1,108 +1,0 @@
-// ***********************************************************************
-// Copyright (c) 2007 Charlie Poole
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// ***********************************************************************
-
-using System;
-using System.Collections.Generic;
-using NUnit.Framework.Extensibility;
-using NUnit.Framework.Internal.Builders;
-
-namespace NUnit.Framework.Internal.Extensibility
-{
-	/// <summary>
-	/// SuiteBuilderCollection is an ExtensionPoint for SuiteBuilders and
-	/// implements the ISuiteBuilder interface itself, passing calls 
-	/// on to the individual builders.
-	/// 
-	/// The builders are added to the collection by inserting them at
-	/// the start, as to take precedence over those added earlier. 
-	/// </summary>
-#if NUNITLITE
-    public class SuiteBuilderCollection : ISuiteBuilder
-    {
-        private List<ISuiteBuilder> Extensions = new List<ISuiteBuilder>();
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SuiteBuilderCollection"/> class.
-        /// </summary>
-        public SuiteBuilderCollection()
-        {
-            Extensions.Add(new NUnitTestFixtureBuilder());
-            Extensions.Add(new SetUpFixtureBuilder());
-        }
-#else
-	public class SuiteBuilderCollection : ExtensionPoint, ISuiteBuilder
-	{
-		#region Constructor
-		/// <summary>
-		/// Default constructor
-		/// </summary>
-		public SuiteBuilderCollection(IExtensionHost host)
-			: base("SuiteBuilders", host ) { }
-		#endregion
-#endif
-
-		#region ISuiteBuilder Members
-
-		/// <summary>
-		/// Examine the type and determine if it is suitable for
-		/// any SuiteBuilder to use in building a TestSuite
-		/// </summary>
-		/// <param name="type">The type of the fixture to be used</param>
-		/// <returns>True if the type can be used to build a TestSuite</returns>
-		public bool CanBuildFrom(Type type)
-		{
-			foreach( ISuiteBuilder builder in Extensions )
-				if ( builder.CanBuildFrom( type ) )
-					return true;
-			return false;
-		}
-
-		/// <summary>
-		/// Build a TestSuite from type provided.
-		/// </summary>
-		/// <param name="type">The type of the fixture to be used</param>
-		/// <returns>A TestSuite or null</returns>
-		public Test BuildFrom(Type type)
-		{
-			foreach( ISuiteBuilder builder in Extensions )
-				if ( builder.CanBuildFrom( type ) )
-					return builder.BuildFrom( type );
-			return null;
-		}
-
-		#endregion
-
-#if !NUNITLITE
-		#region ExtensionPoint Overrides
-        /// <summary>
-        /// Determines whether the specified extension is an ISuiteBuilder.
-        /// </summary>
-        /// <param name="extension">The extension.</param>
-		protected override bool IsValidExtension(object extension)
-		{
-			return extension is ISuiteBuilder; 
-		}
-		#endregion
-#endif
-	}
-}
