+++ src/framework/nunit.framework-4.0.csproj
@@ -79,7 +79,6 @@
     <Compile Include="Attributes\RandomAttribute.cs" />
     <Compile Include="Attributes\RangeAttribute.cs" />
     <Compile Include="Attributes\RepeatAttribute.cs" />
-    <Compile Include="Attributes\RequiredAddinAttribute.cs" />
     <Compile Include="Attributes\RequiresMTAAtribute.cs" />
     <Compile Include="Attributes\RequiresSTAAttribute.cs" />
     <Compile Include="Attributes\RequiresThreadAttribute.cs" />
@@ -202,19 +201,7 @@
     <Compile Include="Exceptions\IgnoreException.cs" />
     <Compile Include="Exceptions\InconclusiveException.cs" />
     <Compile Include="Exceptions\SuccessException.cs" />
-    <Compile Include="Extensibility\Addin.cs" />
-    <Compile Include="Extensibility\AddinStatus.cs" />
-    <Compile Include="Extensibility\ExtensionType.cs" />
-    <Compile Include="Extensibility\IAddin.cs" />
-    <Compile Include="Extensibility\IAddinRegistry.cs" />
-    <Compile Include="Extensibility\IExtensionHost.cs" />
-    <Compile Include="Extensibility\IExtensionPoint.cs" />
     <Compile Include="Interfaces\IParameterDataProvider.cs" />
-    <Compile Include="Extensibility\ISuiteBuilder.cs" />
-    <Compile Include="Extensibility\ITestCaseBuilder.cs" />
-    <Compile Include="Extensibility\ITestCaseProvider.cs" />
-    <Compile Include="Extensibility\NUnitAddinAttribute.cs" />
-    <Compile Include="Extensibility\TestFramework.cs" />
     <Compile Include="FileAssert.cs" />
     <Compile Include="GlobalSettings.cs" />
     <Compile Include="Guard.cs" />
@@ -223,10 +210,13 @@
     <Compile Include="Interfaces\ExpectedExceptionData.cs" />
     <Compile Include="Interfaces\IParameterDataSource.cs" />
     <Compile Include="Interfaces\IPropertyBag.cs" />
+    <Compile Include="Interfaces\ISuiteBuilder.cs" />
     <Compile Include="Interfaces\ITest.cs" />
     <Compile Include="Interfaces\ITestAssemblyBuilder.cs" />
     <Compile Include="Interfaces\ITestAssemblyRunner.cs" />
+    <Compile Include="Interfaces\ITestCaseBuilder.cs" />
     <Compile Include="Interfaces\ITestCaseData.cs" />
+    <Compile Include="Interfaces\ITestCaseProvider.cs" />
     <Compile Include="Interfaces\ITestCaseSource.cs" />
     <Compile Include="Interfaces\ITestCaseSourceProvider.cs" />
     <Compile Include="Interfaces\ITestFilter.cs" />
@@ -255,6 +245,9 @@
     <Compile Include="Internal\Builders\ProviderCache.cs" />
     <Compile Include="Internal\Builders\SequentialStrategy.cs" />
     <Compile Include="Internal\Builders\SetUpFixtureBuilder.cs" />
+    <Compile Include="Internal\Builders\SuiteBuilderCollection.cs" />
+    <Compile Include="Internal\Builders\TestCaseBuilderCollection.cs" />
+    <Compile Include="Internal\Builders\TestCaseProviders.cs" />
     <Compile Include="Internal\Commands\ApplyChangesToContextCommand.cs" />
     <Compile Include="Internal\Commands\CommandDecoratorList.cs" />
     <Compile Include="Internal\Commands\CommandStage.cs" />
@@ -270,7 +263,6 @@
     <Compile Include="Internal\Commands\TestCommand.cs" />
     <Compile Include="Internal\Commands\TestMethodCommand.cs" />
     <Compile Include="Internal\Commands\TheoryResultCommand.cs" />
-    <Compile Include="Internal\CoreExtensions.cs" />
     <Compile Include="Internal\CultureDetector.cs" />
     <Compile Include="Internal\DefaultTestAssemblyBuilder.cs" />
     <Compile Include="Internal\DefaultTestAssemblyRunner.cs" />
@@ -291,14 +283,6 @@
     <Compile Include="Internal\Execution\WorkItemDispatcher.cs" />
     <Compile Include="Internal\Execution\WorkItemQueue.cs" />
     <Compile Include="Internal\Execution\WorkItemState.cs" />
-    <Compile Include="Internal\Extensibility\AddinRegistry.cs" />
-    <Compile Include="Internal\Extensibility\EventListenerCollection.cs" />
-    <Compile Include="Internal\Extensibility\ExtensionHost.cs" />
-    <Compile Include="Internal\Extensibility\ExtensionPoint.cs" />
-    <Compile Include="Internal\Extensibility\ExtensionsCollection.cs" />
-    <Compile Include="Internal\Extensibility\SuiteBuilderCollection.cs" />
-    <Compile Include="Internal\Extensibility\TestCaseBuilderCollection.cs" />
-    <Compile Include="Internal\Extensibility\TestCaseProviders.cs" />
     <Compile Include="Internal\Filters\AndFilter.cs" />
     <Compile Include="Internal\Filters\CategoryExpression.cs" />
     <Compile Include="Internal\Filters\CategoryFilter.cs" />
