+++ src/tests/nunit.framework.tests-2.0.csproj
@@ -159,7 +159,6 @@
     <Compile Include="Constraints\XmlSerializableTest.cs" />
     <Compile Include="Internal\AssemblyHelperTests.cs" />
     <Compile Include="Internal\CallContextTests.cs" />
-    <Compile Include="Internal\CoreExtensionsTests.cs" />
     <Compile Include="Internal\CultureSettingAndDetectionTests.cs" />
     <Compile Include="Internal\DeduceTypeArgsFromArgs.cs" />
     <Compile Include="Internal\EventQueueTests.cs" />
@@ -235,5 +234,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="nunit.framework.tests.build" />
+    <None Include="nunitlite.tests.build" />
   </ItemGroup>
 </Project>
